requirements.txt
flask
mysql-connector-python
pandas
sklearn
numpy